requirements.txt
# PDF Processing
pdfplumber>=0.10.0
PyMuPDF>=1.23.0

# Data Processing & Visualization
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# LangChain & RAG
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10

# Supabase (Vector Store)
supabase>=2.0.0

# OpenAI
openai>=1.0.0
tiktoken>=0.5.0

# Web Interface
streamlit>=1.30.0

# Fine-tuning (Part C)
transformers>=4.35.0
datasets>=2.15.0
torch>=2.0.0
scikit-learn>=1.3.0
accelerate>=0.25.0

# Utilities
python-dotenv>=1.0.0
